requirements.txt
flask
flask-sqlalchemy
flask-migrate
python-dotenv
requests
basilica
tweepy
scikit-learn
psycopg2-binary
gunicorn